+++ .github/workflows/linux_ci.yml
@@ -1,6 +1,7 @@
 name: Linux CI
 
 on:
+  workflow_dispatch:
   push:
     branches: [ master ]
   pull_request:
