+++ .github/workflows/windows_ci.yml
@@ -1,6 +1,7 @@
 name: Windows CI
 
 on:
+  workflow_dispatch:
   push:
     branches: [ master ]
 #  pull_request:
